requirements.txt
streamlit==1.28.2

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2

torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

transformers==4.36.2
sentence-transformers==2.5.1
datasets==2.16.1

PyPDF2==3.0.1
python-docx==0.8.11
pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.16.3
python-pptx==0.6.21

plotly==5.18.0
serpapi==0.1.5
joblib==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.1
requests==2.31.0
